backend/requirements.txt
flask
flask-cors
python-dotenv
pymysql
cryptography
bcrypt
